75% done punctuation
Changed how the definition choices work for anagrams. If two was pressed
it gave you the definitions anyway. I didn't touch the other section so
it's up to you if you want it changed.

Made code to choose which section of punctuation the question will give
(none more than twice, and all at least twice).

Made code to select which question to say in the specific section.

Todo:

Made section that asks the question it selects and records/checks their
answer.

Still need more scolon's, apos, and ques to start testing.